fix(repo): PR suggestions
