+++ cli/auth/handlers_users_tui.go
@@ -2,12 +2,9 @@ package auth
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"os"
 	"sort"
 	"strings"
-	"time"
 
 	"github.com/charmbracelet/bubbles/spinner"
 	tea "github.com/charmbracelet/bubbletea"
@@ -17,870 +14,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-const (
-	keyCtrlC    = "ctrl+c"
-	keyEnter    = "enter"
-	keyDown     = "down"
-	roleUser    = "user"
-	roleAdmin   = "admin"
-	sortName    = "name"
-	sortEmail   = "email"
-	sortRole    = "role"
-	sortCreated = "created"
-)
-
-// runGenerateJSON handles JSON mode for key generation
-func runGenerateJSON(ctx context.Context, cmd *cobra.Command, client *Client) error {
-	log := logger.FromContext(ctx)
-
-	// Parse flags
-	name, err := cmd.Flags().GetString("name")
-	if err != nil {
-		return fmt.Errorf("failed to get name flag: %w", err)
-	}
-	description, err := cmd.Flags().GetString("description")
-	if err != nil {
-		return fmt.Errorf("failed to get description flag: %w", err)
-	}
-	expiresStr, err := cmd.Flags().GetString("expires")
-	if err != nil {
-		return fmt.Errorf("failed to get expires flag: %w", err)
-	}
-
-	// Validate expiration date if provided
-	var expires *time.Time
-	if expiresStr != "" {
-		t, err := time.Parse("2006-01-02", expiresStr)
-		if err != nil {
-			return outputJSONError("invalid expiration date format, use YYYY-MM-DD")
-		}
-		expires = &t
-	}
-
-	log.Debug("generating API key in JSON mode",
-		"name", name,
-		"description", description,
-		"expires", expiresStr)
-
-	// Generate the key
-	req := &GenerateKeyRequest{
-		Name:        name,
-		Description: description,
-	}
-	if expires != nil {
-		req.Expires = expires.Format("2006-01-02")
-	}
-
-	apiKey, err := client.GenerateKey(ctx, req)
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to generate API key: %v", err))
-	}
-
-	// Prepare response
-	response := map[string]any{
-		"api_key": apiKey,
-		"created": time.Now().Format(time.RFC3339),
-	}
-
-	if name != "" {
-		response["name"] = name
-	}
-	if description != "" {
-		response["description"] = description
-	}
-	if expires != nil {
-		response["expires"] = expires.Format(time.RFC3339)
-	}
-
-	// Output JSON
-	encoder := json.NewEncoder(os.Stdout)
-	encoder.SetIndent("", "  ")
-	if err := encoder.Encode(response); err != nil {
-		return fmt.Errorf("failed to encode JSON response: %w", err)
-	}
-
-	return nil
-}
-
-// outputJSONError outputs an error in JSON format
-func outputJSONError(message string) error {
-	response := map[string]any{
-		"error": message,
-	}
-
-	encoder := json.NewEncoder(os.Stderr)
-	encoder.SetIndent("", "  ")
-	if err := encoder.Encode(response); err != nil {
-		// If we can't encode the error response, just return the original error
-		return fmt.Errorf("%s", message)
-	}
-
-	return fmt.Errorf("%s", message)
-}
-
-// runGenerateTUI handles TUI mode for key generation
-func runGenerateTUI(ctx context.Context, cmd *cobra.Command, client *Client) error {
-	log := logger.FromContext(ctx)
-
-	// Parse flags for initial values
-	name, err := cmd.Flags().GetString("name")
-	if err != nil {
-		return fmt.Errorf("failed to get name flag: %w", err)
-	}
-	description, err := cmd.Flags().GetString("description")
-	if err != nil {
-		return fmt.Errorf("failed to get description flag: %w", err)
-	}
-	expiresStr, err := cmd.Flags().GetString("expires")
-	if err != nil {
-		return fmt.Errorf("failed to get expires flag: %w", err)
-	}
-
-	log.Debug("generating API key in TUI mode")
-
-	// Create and run the TUI model
-	m := newGenerateModel(ctx, client, name, description, expiresStr)
-	p := tea.NewProgram(&m)
-
-	finalModel, err := p.Run()
-	if err != nil {
-		return fmt.Errorf("failed to run TUI: %w", err)
-	}
-
-	// Check if generation was successful
-	if model, ok := finalModel.(*generateModel); ok {
-		if model.err != nil {
-			return model.err
-		}
-		if !model.generated {
-			return fmt.Errorf("key generation canceled")
-		}
-	}
-
-	return nil
-}
-
-// generateModel represents the TUI model for key generation
-type generateModel struct {
-	ctx         context.Context
-	client      *Client
-	spinner     spinner.Model
-	generating  bool
-	generated   bool
-	apiKey      string
-	name        string
-	description string
-	expires     string
-	err         error
-	width       int
-	height      int
-}
-
-// newGenerateModel creates a new TUI model for key generation
-func newGenerateModel(ctx context.Context, client *Client, name, description, expires string) generateModel {
-	s := spinner.New()
-	s.Spinner = spinner.Dot
-	s.Style = lipgloss.NewStyle().Foreground(lipgloss.Color("69"))
-
-	return generateModel{
-		ctx:         ctx,
-		client:      client,
-		spinner:     s,
-		name:        name,
-		description: description,
-		expires:     expires,
-	}
-}
-
-// Init initializes the model
-func (m *generateModel) Init() tea.Cmd {
-	return tea.Batch(
-		m.spinner.Tick,
-		m.generateKey(),
-	)
-}
-
-// Update handles messages
-func (m *generateModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tea.WindowSizeMsg:
-		m.width = msg.Width
-		m.height = msg.Height
-
-	case tea.KeyMsg:
-		switch msg.String() {
-		case "q", keyCtrlC:
-			return m, tea.Quit
-		case keyEnter:
-			if m.generated {
-				return m, tea.Quit
-			}
-		}
-
-	case spinner.TickMsg:
-		var cmd tea.Cmd
-		m.spinner, cmd = m.spinner.Update(msg)
-		return m, cmd
-
-	case keyGeneratedMsg:
-		m.generating = false
-		m.generated = true
-		m.apiKey = string(msg)
-		return m, nil
-
-	case errMsg:
-		m.generating = false
-		m.err = msg.err
-		return m, tea.Quit
-	}
-
-	return m, nil
-}
-
-// View renders the UI
-func (m *generateModel) View() string {
-	if m.err != nil {
-		return lipgloss.NewStyle().
-			Foreground(lipgloss.Color("196")).
-			Render(fmt.Sprintf("❌ Error: %v", m.err))
-	}
-
-	if m.generating {
-		return fmt.Sprintf("%s Generating API key...", m.spinner.View())
-	}
-
-	if m.generated {
-		style := lipgloss.NewStyle().
-			Border(lipgloss.RoundedBorder()).
-			BorderForeground(lipgloss.Color("69")).
-			Padding(1, 2).
-			Width(60)
-
-		titleStyle := lipgloss.NewStyle().
-			Bold(true).
-			Foreground(lipgloss.Color("69"))
-
-		keyStyle := lipgloss.NewStyle().
-			Foreground(lipgloss.Color("214")).
-			Bold(true)
-
-		content := titleStyle.Render("✅ API Key Generated Successfully!") + "\n\n"
-		content += "Your new API key:\n"
-		content += keyStyle.Render(m.apiKey) + "\n\n"
-
-		if m.name != "" {
-			content += fmt.Sprintf("Name: %s\n", m.name)
-		}
-		if m.description != "" {
-			content += fmt.Sprintf("Description: %s\n", m.description)
-		}
-		if m.expires != "" {
-			content += fmt.Sprintf("Expires: %s\n", m.expires)
-		}
-
-		content += "\n" + lipgloss.NewStyle().
-			Foreground(lipgloss.Color("241")).
-			Render("⚠️  Save this key securely. You won't be able to see it again!")
-
-		content += "\n\nPress Enter or 'q' to exit"
-
-		return style.Render(content)
-	}
-
-	return ""
-}
-
-// generateKey generates the API key
-func (m *generateModel) generateKey() tea.Cmd {
-	return func() tea.Msg {
-		req := &GenerateKeyRequest{
-			Name:        m.name,
-			Description: m.description,
-			Expires:     m.expires,
-		}
-		apiKey, err := m.client.GenerateKey(m.ctx, req)
-		if err != nil {
-			return errMsg{err}
-		}
-		return keyGeneratedMsg(apiKey)
-	}
-}
-
-// Message types for the TUI
-type keyGeneratedMsg string
-type errMsg struct{ err error }
-
-// runListJSON handles JSON mode for key listing
-func runListJSON(ctx context.Context, cmd *cobra.Command, client *Client) error {
-	log := logger.FromContext(ctx)
-	// Parse flags
-	sortBy, err := cmd.Flags().GetString("sort")
-	if err != nil {
-		return fmt.Errorf("failed to get sort flag: %w", err)
-	}
-	filter, err := cmd.Flags().GetString("filter")
-	if err != nil {
-		return fmt.Errorf("failed to get filter flag: %w", err)
-	}
-	page, err := cmd.Flags().GetInt("page")
-	if err != nil {
-		return fmt.Errorf("failed to get page flag: %w", err)
-	}
-	limit, err := cmd.Flags().GetInt("limit")
-	if err != nil {
-		return fmt.Errorf("failed to get limit flag: %w", err)
-	}
-	log.Debug("listing API keys in JSON mode",
-		"sort", sortBy,
-		"filter", filter,
-		"page", page,
-		"limit", limit)
-	// Get the keys from the API
-	keys, err := client.ListKeys(ctx)
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to list API keys: %v", err))
-	}
-	// Apply client-side filtering if needed
-	if filter != "" {
-		filtered := make([]models.KeyInfo, 0)
-		for _, key := range keys {
-			// Filter by prefix or ID
-			if contains(key.Prefix, filter) || contains(key.ID, filter) {
-				filtered = append(filtered, key)
-			}
-		}
-		keys = filtered
-	}
-	// Apply client-side sorting
-	sortKeys(keys, sortBy)
-	// Apply client-side pagination
-	totalKeys := len(keys)
-	startIdx := (page - 1) * limit
-	endIdx := startIdx + limit
-	if startIdx >= totalKeys {
-		// Return empty result if page is out of bounds
-		keys = []models.KeyInfo{}
-	} else {
-		if endIdx > totalKeys {
-			endIdx = totalKeys
-		}
-		keys = keys[startIdx:endIdx]
-	}
-	// Prepare response with pagination metadata
-	response := map[string]any{
-		"keys":  keys,
-		"total": totalKeys,
-		"page":  page,
-		"limit": limit,
-		"pages": (totalKeys + limit - 1) / limit,
-	}
-	// Output JSON
-	encoder := json.NewEncoder(os.Stdout)
-	encoder.SetIndent("", "  ")
-	if err := encoder.Encode(response); err != nil {
-		return fmt.Errorf("failed to encode JSON response: %w", err)
-	}
-	return nil
-}
-
-// runListTUI handles TUI mode for key listing
-func runListTUI(ctx context.Context, _ *cobra.Command, client *Client) error {
-	log := logger.FromContext(ctx)
-	log.Debug("listing API keys in TUI mode")
-	// Create and run the TUI model
-	m := models.NewListModel(ctx, client)
-	p := tea.NewProgram(m)
-	finalModel, err := p.Run()
-	if err != nil {
-		return fmt.Errorf("failed to run TUI: %w", err)
-	}
-	// Check if there was an error
-	if model, ok := finalModel.(*models.ListModel); ok {
-		if model.Error() != nil {
-			return model.Error()
-		}
-	}
-	return nil
-}
-
-// contains checks if a string contains a substring (case-insensitive)
-func contains(s, substr string) bool {
-	return strings.Contains(strings.ToLower(s), strings.ToLower(substr))
-}
-
-// sortKeys sorts a slice of KeyInfo based on the specified field
-func sortKeys(keys []models.KeyInfo, sortBy string) {
-	switch sortBy {
-	case sortName:
-		// Sort by prefix (which contains the name prefix)
-		sort.Slice(keys, func(i, j int) bool {
-			return keys[i].Prefix < keys[j].Prefix
-		})
-	case "last_used":
-		// Sort by last used timestamp (most recent first)
-		sort.Slice(keys, func(i, j int) bool {
-			// Handle nil LastUsed values
-			if keys[i].LastUsed == nil && keys[j].LastUsed == nil {
-				return false
-			}
-			if keys[i].LastUsed == nil {
-				return false
-			}
-			if keys[j].LastUsed == nil {
-				return true
-			}
-			return *keys[i].LastUsed > *keys[j].LastUsed
-		})
-	case sortCreated, "":
-		// Sort by created timestamp (most recent first)
-		sort.Slice(keys, func(i, j int) bool {
-			return keys[i].CreatedAt > keys[j].CreatedAt
-		})
-	default:
-		// Sort by created timestamp (most recent first) for any unrecognized sort field
-		sort.Slice(keys, func(i, j int) bool {
-			return keys[i].CreatedAt > keys[j].CreatedAt
-		})
-	}
-}
-
-// runRevokeJSON handles JSON mode for key revocation
-func runRevokeJSON(ctx context.Context, cmd *cobra.Command, client *Client, keyID string) error {
-	log := logger.FromContext(ctx)
-
-	// Get force flag
-	force, err := cmd.Flags().GetBool("force")
-	if err != nil {
-		return fmt.Errorf("failed to get force flag: %w", err)
-	}
-
-	log.Debug("revoking API key in JSON mode",
-		"key_id", keyID,
-		"force", force)
-
-	// If not forced, we should show a warning (in a real implementation,
-	// we'd show affected resources)
-	if !force {
-		// For now, just show a confirmation prompt via stderr
-		fmt.Fprintf(os.Stderr, "Warning: This will permanently revoke the API key.\n")
-		fmt.Fprintf(os.Stderr, "Use --force to skip this confirmation.\n")
-		fmt.Fprintf(os.Stderr, "Continue? (y/N): ")
-
-		var response string
-		if _, err := fmt.Scanln(&response); err != nil {
-			return outputJSONError("revocation canceled")
-		}
-		if response != "y" && response != "Y" {
-			return outputJSONError("revocation canceled")
-		}
-	}
-
-	// Revoke the key
-	err = client.RevokeKey(ctx, keyID)
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to revoke API key: %v", err))
-	}
-
-	// Prepare response
-	response := map[string]any{
-		"message": "API key revoked successfully",
-		"key_id":  keyID,
-		"revoked": time.Now().Format(time.RFC3339),
-	}
-
-	// Output JSON
-	encoder := json.NewEncoder(os.Stdout)
-	encoder.SetIndent("", "  ")
-	if err := encoder.Encode(response); err != nil {
-		return fmt.Errorf("failed to encode JSON response: %w", err)
-	}
-
-	return nil
-}
-
-// runRevokeTUI handles TUI mode for key revocation
-func runRevokeTUI(ctx context.Context, _ *cobra.Command, client *Client) error {
-	log := logger.FromContext(ctx)
-	log.Debug("revoking API key in TUI mode")
-
-	// Create and run the TUI model
-	m := newRevokeModel(ctx, client)
-	p := tea.NewProgram(m)
-
-	finalModel, err := p.Run()
-	if err != nil {
-		return fmt.Errorf("failed to run TUI: %w", err)
-	}
-
-	// Check if revocation was successful
-	if model, ok := finalModel.(*revokeModel); ok {
-		if model.err != nil {
-			return model.err
-		}
-		if !model.revoked {
-			return fmt.Errorf("key revocation canceled")
-		}
-	}
-
-	return nil
-}
-
-// revokeModel represents the TUI model for key revocation
-type revokeModel struct {
-	ctx      context.Context
-	client   *Client
-	keys     []models.KeyInfo
-	selected int
-	state    revokeState
-	revoked  bool
-	err      error
-	width    int
-	height   int
-	spinner  spinner.Model
-	loading  bool
-}
-
-type revokeState int
-
-const (
-	stateLoadingKeys revokeState = iota
-	stateSelectingKey
-	stateConfirming
-	stateRevoking
-	stateDone
-)
-
-// newRevokeModel creates a new TUI model for key revocation
-func newRevokeModel(ctx context.Context, client *Client) *revokeModel {
-	s := spinner.New()
-	s.Spinner = spinner.Dot
-	s.Style = lipgloss.NewStyle().Foreground(lipgloss.Color("69"))
-
-	return &revokeModel{
-		ctx:     ctx,
-		client:  client,
-		state:   stateLoadingKeys,
-		spinner: s,
-		loading: true,
-	}
-}
-
-// Init initializes the model
-func (m *revokeModel) Init() tea.Cmd {
-	return tea.Batch(
-		m.spinner.Tick,
-		m.loadKeys(),
-	)
-}
-
-// Update handles messages
-func (m *revokeModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tea.WindowSizeMsg:
-		m.width = msg.Width
-		m.height = msg.Height
-
-	case tea.KeyMsg:
-		return m.handleKeyMsg(msg)
-
-	case revokeKeysLoadedMsg:
-		m.loading = false
-		m.keys = msg.keys
-		if len(m.keys) == 0 {
-			m.err = fmt.Errorf("no API keys found")
-			return m, tea.Quit
-		}
-		m.state = stateSelectingKey
-		return m, nil
-
-	case revokeKeyRevokedMsg:
-		m.loading = false
-		m.revoked = true
-		m.state = stateDone
-		return m, nil
-
-	case errMsg:
-		m.loading = false
-		m.err = msg.err
-		return m, tea.Quit
-
-	case spinner.TickMsg:
-		if m.loading {
-			var cmd tea.Cmd
-			m.spinner, cmd = m.spinner.Update(msg)
-			return m, cmd
-		}
-	}
-
-	return m, nil
-}
-
-// handleKeyMsg handles keyboard input
-func (m *revokeModel) handleKeyMsg(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch m.state {
-	case stateSelectingKey:
-		return m.handleSelectingKeyInput(msg)
-	case stateConfirming:
-		return m.handleConfirmingInput(msg)
-	case stateDone:
-		return m, tea.Quit
-	}
-	return m, nil
-}
-
-// handleSelectingKeyInput handles keyboard input in the selecting key state
-func (m *revokeModel) handleSelectingKeyInput(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "q", keyCtrlC:
-		return m, tea.Quit
-	case "up", "k":
-		if m.selected > 0 {
-			m.selected--
-		}
-	case keyDown, "j":
-		if m.selected < len(m.keys)-1 {
-			m.selected++
-		}
-	case keyEnter:
-		m.state = stateConfirming
-		return m, nil
-	}
-	return m, nil
-}
-
-// handleConfirmingInput handles keyboard input in the confirming state
-func (m *revokeModel) handleConfirmingInput(msg tea.KeyMsg) (tea.Model, tea.Cmd) {
-	switch msg.String() {
-	case "y", "Y":
-		m.state = stateRevoking
-		cmd := m.revokeKey()
-		return m, cmd
-	case "n", "N", "q", keyCtrlC:
-		m.state = stateSelectingKey
-		return m, nil
-	}
-	return m, nil
-}
-
-// View renders the UI
-func (m *revokeModel) View() string {
-	if m.err != nil {
-		return lipgloss.NewStyle().
-			Foreground(lipgloss.Color("196")).
-			Render(fmt.Sprintf("❌ Error: %v", m.err))
-	}
-
-	switch m.state {
-	case stateLoadingKeys:
-		return fmt.Sprintf("%s Loading API keys...", m.spinner.View())
-	case stateSelectingKey:
-		return m.viewKeySelection()
-	case stateConfirming:
-		return m.viewConfirmation()
-	case stateRevoking:
-		return fmt.Sprintf("%s Revoking API key...", m.spinner.View())
-	case stateDone:
-		return m.viewDone()
-	}
-
-	return ""
-}
-
-// viewKeySelection renders the key selection view
-func (m *revokeModel) viewKeySelection() string {
-	style := lipgloss.NewStyle().
-		Border(lipgloss.RoundedBorder()).
-		BorderForeground(lipgloss.Color("69")).
-		Padding(1, 2).
-		Width(80)
-
-	titleStyle := lipgloss.NewStyle().
-		Bold(true).
-		Foreground(lipgloss.Color("69"))
-
-	selectedStyle := lipgloss.NewStyle().
-		Foreground(lipgloss.Color("229")).
-		Background(lipgloss.Color("57")).
-		Bold(false)
-
-	content := titleStyle.Render("🔑 Select API Key to Revoke") + "\n\n"
-
-	for i, key := range m.keys {
-		created := key.CreatedAt
-		if t, err := time.Parse(time.RFC3339, key.CreatedAt); err == nil {
-			created = t.Format("2006-01-02 15:04")
-		}
-
-		line := fmt.Sprintf("  %s - Created: %s", key.Prefix, created)
-		if key.LastUsed != nil {
-			if t, err := time.Parse(time.RFC3339, *key.LastUsed); err == nil {
-				line += fmt.Sprintf(" - Last used: %s", t.Format("2006-01-02 15:04"))
-			}
-		}
-
-		if i == m.selected {
-			content += selectedStyle.Render("> "+line[2:]) + "\n"
-		} else {
-			content += line + "\n"
-		}
-	}
-
-	content += "\n" + lipgloss.NewStyle().
-		Foreground(lipgloss.Color("241")).
-		Render("↑/↓ to navigate • Enter to select • q to quit")
-
-	return style.Render(content)
-}
-
-// viewConfirmation renders the confirmation view
-func (m *revokeModel) viewConfirmation() string {
-	style := lipgloss.NewStyle().
-		Border(lipgloss.RoundedBorder()).
-		BorderForeground(lipgloss.Color("214")).
-		Padding(1, 2).
-		Width(60)
-
-	titleStyle := lipgloss.NewStyle().
-		Bold(true).
-		Foreground(lipgloss.Color("214"))
-
-	warningStyle := lipgloss.NewStyle().
-		Foreground(lipgloss.Color("196")).
-		Bold(true)
-
-	keyStyle := lipgloss.NewStyle().
-		Foreground(lipgloss.Color("69"))
-
-	selectedKey := m.keys[m.selected]
-	content := titleStyle.Render("⚠️  Confirm Key Revocation") + "\n\n"
-	content += "You are about to revoke the following API key:\n\n"
-	content += keyStyle.Render(fmt.Sprintf("  Key: %s", selectedKey.Prefix)) + "\n"
-	content += fmt.Sprintf("  ID: %s\n", selectedKey.ID)
-
-	created := selectedKey.CreatedAt
-	if t, err := time.Parse(time.RFC3339, selectedKey.CreatedAt); err == nil {
-		created = t.Format("2006-01-02 15:04")
-	}
-	content += fmt.Sprintf("  Created: %s\n", created)
-
-	if selectedKey.LastUsed != nil {
-		if t, err := time.Parse(time.RFC3339, *selectedKey.LastUsed); err == nil {
-			content += fmt.Sprintf("  Last used: %s\n", t.Format("2006-01-02 15:04"))
-		}
-	}
-
-	content += "\n" + warningStyle.Render("This action cannot be undone!") + "\n\n"
-	content += "Are you sure you want to revoke this key? (y/N)"
-
-	return style.Render(content)
-}
-
-// viewDone renders the completion view
-func (m *revokeModel) viewDone() string {
-	style := lipgloss.NewStyle().
-		Border(lipgloss.RoundedBorder()).
-		BorderForeground(lipgloss.Color("69")).
-		Padding(1, 2).
-		Width(60)
-
-	titleStyle := lipgloss.NewStyle().
-		Bold(true).
-		Foreground(lipgloss.Color("69"))
-
-	content := titleStyle.Render("✅ API Key Revoked Successfully!") + "\n\n"
-	content += fmt.Sprintf("The key %s has been revoked.\n", m.keys[m.selected].Prefix)
-	content += "\nPress any key to exit"
-
-	return style.Render(content)
-}
-
-// loadKeys loads the API keys
-func (m *revokeModel) loadKeys() tea.Cmd {
-	return func() tea.Msg {
-		keys, err := m.client.ListKeys(m.ctx)
-		if err != nil {
-			return errMsg{err}
-		}
-		return revokeKeysLoadedMsg{keys: keys}
-	}
-}
-
-// revokeKey revokes the selected key
-func (m *revokeModel) revokeKey() tea.Cmd {
-	return func() tea.Msg {
-		selectedKey := m.keys[m.selected]
-		err := m.client.RevokeKey(m.ctx, selectedKey.ID)
-		if err != nil {
-			return errMsg{err}
-		}
-		return revokeKeyRevokedMsg{}
-	}
-}
-
-// Message types for the revoke TUI
-type revokeKeysLoadedMsg struct {
-	keys []models.KeyInfo
-}
-
-type revokeKeyRevokedMsg struct{}
-
-// runCreateUserJSON handles JSON mode for user creation
-func runCreateUserJSON(ctx context.Context, cmd *cobra.Command, client *Client) error {
-	log := logger.FromContext(ctx)
-
-	// Parse flags
-	email, err := cmd.Flags().GetString("email")
-	if err != nil {
-		return fmt.Errorf("failed to get email flag: %w", err)
-	}
-	name, err := cmd.Flags().GetString("name")
-	if err != nil {
-		return fmt.Errorf("failed to get name flag: %w", err)
-	}
-	role, err := cmd.Flags().GetString("role")
-	if err != nil {
-		return fmt.Errorf("failed to get role flag: %w", err)
-	}
-
-	// Validate role
-	if role != roleAdmin && role != roleUser {
-		return outputJSONError("invalid role: must be 'admin' or 'user'")
-	}
-
-	log.Debug("creating user in JSON mode",
-		"email", email,
-		"name", name,
-		"role", role)
-
-	// Create the user
-	req := CreateUserRequest{
-		Email: email,
-		Name:  name,
-		Role:  role,
-	}
-
-	user, err := client.CreateUser(ctx, req)
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to create user: %v", err))
-	}
-
-	// Prepare response
-	response := map[string]any{
-		"user":    user,
-		"message": "User created successfully",
-	}
-
-	// Output JSON
-	encoder := json.NewEncoder(os.Stdout)
-	encoder.SetIndent("", "  ")
-	if err := encoder.Encode(response); err != nil {
-		return fmt.Errorf("failed to encode JSON response: %w", err)
-	}
-
-	return nil
-}
-
-// runCreateUserTUI handles TUI mode for user creation
 func runCreateUserTUI(ctx context.Context, cmd *cobra.Command, client *Client) error {
 	log := logger.FromContext(ctx)
 
@@ -1077,7 +210,7 @@ func (m *createUserModel) validateInputs() error {
 		return fmt.Errorf("email is required")
 	}
 	if m.inputs[2] != roleAdmin && m.inputs[2] != roleUser {
-		return fmt.Errorf("role must be 'admin' or 'user'")
+		return fmt.Errorf("role must be '%s' or '%s'", roleAdmin, roleUser)
 	}
 	return nil
 }
@@ -1232,146 +365,6 @@ func (m *createUserModel) createUser() tea.Cmd {
 // Message types for the create user TUI
 type userCreatedMsg struct {
 	user *models.UserInfo
-}
-
-// userFilters holds the parsed command line flags for user filtering
-type userFilters struct {
-	roleFilter string
-	sortBy     string
-	filterStr  string
-	activeOnly bool
-}
-
-// parseListUsersFlags extracts and validates flags for user listing
-func parseListUsersFlags(cmd *cobra.Command) (*userFilters, error) {
-	roleFilter, err := cmd.Flags().GetString("role")
-	if err != nil {
-		return nil, fmt.Errorf("failed to get role flag: %w", err)
-	}
-	sortBy, err := cmd.Flags().GetString("sort")
-	if err != nil {
-		return nil, fmt.Errorf("failed to get sort flag: %w", err)
-	}
-	filterStr, err := cmd.Flags().GetString("filter")
-	if err != nil {
-		return nil, fmt.Errorf("failed to get filter flag: %w", err)
-	}
-	activeOnly, err := cmd.Flags().GetBool("active")
-	if err != nil {
-		return nil, fmt.Errorf("failed to get active flag: %w", err)
-	}
-
-	// Validate role filter
-	if roleFilter != "" && roleFilter != roleAdmin && roleFilter != roleUser {
-		return nil, fmt.Errorf("invalid role filter: %s (must be 'admin' or 'user')", roleFilter)
-	}
-
-	// Validate sort field
-	validSorts := []string{sortCreated, sortName, sortEmail, sortRole}
-	validSort := false
-	for _, valid := range validSorts {
-		if sortBy == valid {
-			validSort = true
-			break
-		}
-	}
-	if !validSort {
-		return nil, fmt.Errorf("invalid sort field: %s (must be one of: %v)", sortBy, validSorts)
-	}
-
-	return &userFilters{
-		roleFilter: roleFilter,
-		sortBy:     sortBy,
-		filterStr:  filterStr,
-		activeOnly: activeOnly,
-	}, nil
-}
-
-// filterAndSortUsers applies filtering and sorting to the user list
-func filterAndSortUsers(users []models.UserInfo, filters *userFilters) []models.UserInfo {
-	filtered := make([]models.UserInfo, 0, len(users))
-
-	for _, user := range users {
-		// Apply role filter
-		if filters.roleFilter != "" && user.Role != filters.roleFilter {
-			continue
-		}
-
-		// Apply text filter (name or email)
-		if filters.filterStr != "" {
-			if !contains(user.Name, filters.filterStr) && !contains(user.Email, filters.filterStr) {
-				continue
-			}
-		}
-
-		// TODO: Apply active filter when KeyCount field is available
-		// For now, we skip the active filter
-		if filters.activeOnly {
-			// Skip for now - would need API key count information
-			continue
-		}
-
-		filtered = append(filtered, user)
-	}
-
-	// Sort users
-	sort.Slice(filtered, func(i, j int) bool {
-		switch filters.sortBy {
-		case sortName:
-			return filtered[i].Name < filtered[j].Name
-		case sortEmail:
-			return filtered[i].Email < filtered[j].Email
-		case sortRole:
-			return filtered[i].Role < filtered[j].Role
-		case sortCreated:
-			return filtered[i].CreatedAt < filtered[j].CreatedAt
-		default:
-			return filtered[i].CreatedAt < filtered[j].CreatedAt
-		}
-	})
-
-	return filtered
-}
-
-// runListUsersJSON handles JSON mode for user listing
-func runListUsersJSON(ctx context.Context, cmd *cobra.Command, client *Client) error {
-	log := logger.FromContext(ctx)
-
-	// Parse flags
-	filters, err := parseListUsersFlags(cmd)
-	if err != nil {
-		return err
-	}
-
-	log.Debug("listing users in JSON mode",
-		"role", filters.roleFilter,
-		"sort", filters.sortBy,
-		"filter", filters.filterStr,
-		"activeOnly", filters.activeOnly)
-
-	// Get users from API
-	users, err := client.ListUsers(ctx)
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to list users: %v", err))
-	}
-
-	// Apply filters and sorting
-	filteredUsers := filterAndSortUsers(users, filters)
-
-	// Prepare response
-	response := map[string]any{
-		"users": filteredUsers,
-		"total": len(filteredUsers),
-	}
-
-	// Output JSON
-	encoder := json.NewEncoder(os.Stdout)
-	encoder.SetIndent("", "  ")
-	if err := encoder.Encode(response); err != nil {
-		return fmt.Errorf("failed to encode JSON response: %w", err)
-	}
-
-	return nil
 }
 
 // runListUsersTUI handles TUI mode for user listing
@@ -1820,69 +813,6 @@ type usersListLoadedMsg struct {
 	users []models.UserInfo
 }
 
-// runUpdateUserJSON handles JSON mode for user updates
-func runUpdateUserJSON(ctx context.Context, cmd *cobra.Command, client *Client, userID string) error {
-	log := logger.FromContext(ctx)
-
-	// Parse flags
-	email, err := cmd.Flags().GetString("email")
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to get email flag: %v", err))
-	}
-	name, err := cmd.Flags().GetString("name")
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to get name flag: %v", err))
-	}
-	role, err := cmd.Flags().GetString("role")
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to get role flag: %v", err))
-	}
-
-	// Validate role if provided
-	if role != "" && role != roleUser && role != roleAdmin {
-		return outputJSONError("role must be 'user' or 'admin'")
-	}
-
-	// Create update request with only specified fields
-	req := UpdateUserRequest{}
-	if email != "" {
-		req.Email = &email
-	}
-	if name != "" {
-		req.Name = &name
-	}
-	if role != "" {
-		req.Role = &role
-	}
-
-	// Check if any fields were provided
-	if req.Email == nil && req.Name == nil && req.Role == nil {
-		return outputJSONError("at least one field (email, name, role) must be provided")
-	}
-
-	log.Debug("updating user in JSON mode", "user_id", userID)
-
-	// Update the user
-	user, err := client.UpdateUser(ctx, userID, req)
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to update user: %v", err))
-	}
-
-	// Output JSON response
-	response := map[string]any{
-		"user":    user,
-		"message": "User updated successfully",
-	}
-
-	encoder := json.NewEncoder(os.Stdout)
-	encoder.SetIndent("", "  ")
-	if err := encoder.Encode(response); err != nil {
-		return fmt.Errorf("failed to encode JSON response: %w", err)
-	}
-
-	return nil
-}
-
 // runUpdateUserTUI handles TUI mode for user updates
 func runUpdateUserTUI(ctx context.Context, _ *cobra.Command, client *Client, userID string) error {
 	log := logger.FromContext(ctx)
@@ -1909,54 +839,6 @@ func runUpdateUserTUI(ctx context.Context, _ *cobra.Command, client *Client, use
 
 	if model.IsQuitting() && !model.success {
 		fmt.Println("Operation canceled")
-	}
-
-	return nil
-}
-
-// runDeleteUserJSON handles JSON mode for user deletion
-func runDeleteUserJSON(ctx context.Context, cmd *cobra.Command, client *Client, userID string) error {
-	log := logger.FromContext(ctx)
-
-	// Parse flags
-	force, err := cmd.Flags().GetBool("force")
-	if err != nil {
-		return fmt.Errorf("failed to get force flag: %w", err)
-	}
-	cascade, err := cmd.Flags().GetBool("cascade")
-	if err != nil {
-		return fmt.Errorf("failed to get cascade flag: %w", err)
-	}
-
-	log.Debug("deleting user in JSON mode", "user_id", userID, "force", force, "cascade", cascade)
-
-	// If not forced, show warning and require confirmation
-	if !force {
-		return outputJSONError("user deletion requires --force flag in JSON mode for safety")
-	}
-
-	// TODO: If cascade is enabled, also delete user's API keys
-	// This would require additional API endpoint or client method
-	if cascade {
-		log.Debug("cascade deletion requested but not yet implemented")
-	}
-
-	// Delete the user
-	err = client.DeleteUser(ctx, userID)
-	if err != nil {
-		return outputJSONError(fmt.Sprintf("failed to delete user: %v", err))
-	}
-
-	// Output JSON response
-	response := map[string]any{
-		"message": "User deleted successfully",
-		"user_id": userID,
-	}
-
-	encoder := json.NewEncoder(os.Stdout)
-	encoder.SetIndent("", "  ")
-	if err := encoder.Encode(response); err != nil {
-		return fmt.Errorf("failed to encode JSON response: %w", err)
 	}
 
 	return nil
