+++ engine/auth/uc/get_api_key.go
@@ -10,21 +10,19 @@ import (
 
 // GetAPIKey use case for retrieving an API key by ID
 type GetAPIKey struct {
-	repo  Repository
-	keyID core.ID
+	repo Repository
 }
 
 // NewGetAPIKey creates a new get API key use case
-func NewGetAPIKey(repo Repository, keyID core.ID) *GetAPIKey {
+func NewGetAPIKey(repo Repository) *GetAPIKey {
 	return &GetAPIKey{
-		repo:  repo,
-		keyID: keyID,
+		repo: repo,
 	}
 }
 
 // Execute retrieves an API key by ID
-func (uc *GetAPIKey) Execute(ctx context.Context) (*model.APIKey, error) {
-	apiKey, err := uc.repo.GetAPIKeyByID(ctx, uc.keyID)
+func (uc *GetAPIKey) Execute(ctx context.Context, keyID core.ID) (*model.APIKey, error) {
+	apiKey, err := uc.repo.GetAPIKeyByID(ctx, keyID)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get API key: %w", err)
 	}
