+++ engine/memory/router/register.go
@@ -18,35 +18,35 @@ func Register(apiBase *gin.RouterGroup, authFactory *uc.Factory) {
 	refGroup := memoryGroup.Group("/:memory_ref")
 	refGroup.Use(ExtractMemoryContext())
 	{
-		// GET /api/v0/memory/:memory_ref/read?key={key}
+		// GET /memory/:memory_ref/read?key={key}
 		// Read memory content
 		refGroup.GET("/read", readMemory)
 
-		// POST /api/v0/memory/:memory_ref/write
+		// POST /memory/:memory_ref/write
 		// Write/replace memory content (key in body)
 		refGroup.POST("/write", writeMemory)
 
-		// POST /api/v0/memory/:memory_ref/append
+		// POST /memory/:memory_ref/append
 		// Append to memory (key in body)
 		refGroup.POST("/append", appendMemory)
 
-		// POST /api/v0/memory/:memory_ref/delete
+		// POST /memory/:memory_ref/delete
 		// Delete memory (key in body)
 		refGroup.POST("/delete", deleteMemory)
 
-		// POST /api/v0/memory/:memory_ref/flush
+		// POST /memory/:memory_ref/flush
 		// Flush memory (key in body)
 		refGroup.POST("/flush", flushMemory)
 
-		// GET /api/v0/memory/:memory_ref/health?key={key}
+		// GET /memory/:memory_ref/health?key={key}
 		// Get memory health
 		refGroup.GET("/health", healthMemory)
 
-		// POST /api/v0/memory/:memory_ref/clear
+		// POST /memory/:memory_ref/clear
 		// Clear memory (key in body)
 		refGroup.POST("/clear", clearMemory)
 
-		// GET /api/v0/memory/:memory_ref/stats?key={key}
+		// GET /memory/:memory_ref/stats?key={key}
 		// Get memory statistics
 		refGroup.GET("/stats", statsMemory)
 	}
